refactor: move crate::add_package::add_package to crate::unity_project::add_package::add_package
